remove unused srx crate
